Refactor files to be under 100 lines each
## File Structure Improvements

### Split Large Files Into Focused Modules
- **utils.py**: 228 lines → 37 lines (moved health checks to separate modules)
- **example_tasks.py**: 329 lines → 15 lines (split into focused task modules)
- **test_utils_redis_celery.py**: 254 lines → 23 lines (split into focused test modules)

### New Modular Structure

#### Health Check Modules (95 lines total)
- `app/api/routes/health/redis.py` - Redis health endpoints (95 lines)
- `app/api/routes/health/celery.py` - Celery health endpoints (43 lines)
- `app/api/routes/health/celery_utils.py` - Celery utilities (77 lines)

#### Task Modules (260 lines total)
- `app/tasks/computation.py` - Heavy computation tasks (75 lines)
- `app/tasks/file_processing.py` - File processing tasks (64 lines)
- `app/tasks/redis_utils.py` - Redis utility tasks (60 lines)
- `app/tasks/redis_client.py` - Redis client utilities (61 lines)
- `app/tasks/file_utils.py` - File processing utilities (29 lines)
- `app/tasks/file_validation.py` - File validation utilities (69 lines)

#### Test Modules (245 lines total)
- `app/tests/api/routes/health/test_redis.py` - Redis health tests (67 lines)
- `app/tests/api/routes/health/test_celery.py` - Celery health tests (55 lines)
- `app/tests/tasks/test_redis_tasks.py` - Redis task tests (30 lines)
- `app/tests/tasks/test_computation_tasks.py` - Computation task tests (40 lines)
- `app/tests/tasks/test_file_tasks.py` - File processing task tests (29 lines)

### Benefits
- **Better Maintainability**: Each file has a single, clear responsibility
- **Easier Testing**: Focused test modules for specific functionality
- **Improved Readability**: Smaller files are easier to understand and navigate
- **Reduced Cognitive Load**: Maximum 100 lines per file guideline enforced
- **Backward Compatibility**: Legacy imports maintained through re-exports

### Updated Configuration
- Updated Celery task discovery to include all new task modules
- Enhanced task routing for proper queue assignment
- Maintained all existing functionality through import aggregation
