+++ backend/app/api/routes/health/__init__.py
@@ -1,0 +1,3 @@
+"""
+Health check endpoints package.
+"""
