+++ backend/app/core/celery.py
@@ -25,21 +25,25 @@ celery_app.conf.update(
     worker_max_tasks_per_child=1000,
     # Auto-discover tasks from all installed apps
     include=[
-        "app.tasks.example_tasks",
+        "app.tasks.computation",
+        "app.tasks.file_processing",
+        "app.tasks.redis_utils",
+        "app.tasks.example_tasks",  # Legacy compatibility
         # Add more task modules here as needed
     ],
 )
 
 # Configure task routes for different queues
 celery_app.conf.task_routes = {
+    "app.tasks.redis_utils.add_numbers": {"queue": "default"},
+    "app.tasks.redis_utils.test_redis_connection": {"queue": "default"},
+    "app.tasks.computation.process_heavy_computation_task": {"queue": "heavy_compute"},
+    "app.tasks.file_processing.process_large_file_upload_task": {"queue": "file_processing"},
+    # Legacy compatibility
     "app.tasks.example_tasks.add_numbers": {"queue": "default"},
     "app.tasks.example_tasks.test_redis_connection": {"queue": "default"},
-    "app.tasks.example_tasks.process_heavy_computation_task": {
-        "queue": "heavy_compute"
-    },
-    "app.tasks.example_tasks.process_large_file_upload_task": {
-        "queue": "file_processing"
-    },
+    "app.tasks.example_tasks.process_heavy_computation_task": {"queue": "heavy_compute"},
+    "app.tasks.example_tasks.process_large_file_upload_task": {"queue": "file_processing"},
     # Add more routing rules as needed
 }
