+++ backend/app/tests/api/routes/health/__init__.py
@@ -1,0 +1,3 @@
+"""
+Health check endpoint tests package.
+"""
