+++ backend/app/tests/tasks/__init__.py
@@ -1,0 +1,3 @@
+"""
+Task tests package.
+"""
